Src/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorama==0.4.6
fastapi==0.115.0
h11==0.16.0
idna==3.12
joblib==1.5.3
numpy==1.26.4
pandas==3.0.2
psycopg2-binary==2.9.10
pydantic==2.9.0
pydantic_core==2.23.2
python-dateutil==2.9.0.post0
scikit-surprise==1.1.4
scipy==1.17.1
six==1.17.0
starlette==0.38.6
typing_extensions==4.15.0
tzdata==2026.1
uvicorn==0.30.0
